web_scanner/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
Pillow>=10.0.0
python-telegram-bot>=20.0
schedule>=1.2.0
stem>=1.8.0
psutil>=5.9.0
sqlite3
lxml>=4.9.0
aiohttp>=3.9.0
numpy>=1.24.0
scikit-learn>=1.3.0